requirements.txt
selenium==4.14.0
pytest==7.4.1
pytest-xdist==3.3.1
requests==2.31.0
allure-pytest==2.13.2
faker==19.10.0